Split test/tests.js into tests/parser.js & tests/messageformat.js
