+++ test/index.html
@@ -9,7 +9,8 @@
     <script src="../node_modules/expect.js/index.js"></script>
     <script>mocha.setup('bdd')</script>
     <script src="../messageformat.js"></script>
-    <script src="tests.js"></script>
+    <script src="parser.js"></script>
+    <script src="messageformat.js"></script>
   </head>
   <body>
     <h1>messageformat.js Test Suite</h1>
